Move Jetpack instructions component to a separate file.
