+++ client/my-sites/checkout/checkout-thank-you/licensing-thank-you-manual-activation-instructions.tsx
@@ -5,118 +5,19 @@ import page from 'page';
 import { FC, useCallback, useMemo } from 'react';
 import { useDispatch } from 'react-redux';
 import licensingActivationPluginInstall from 'calypso/assets/images/jetpack/licensing-activation-plugin-install.svg';
-import ExternalLink from 'calypso/components/external-link';
 import LicensingActivation from 'calypso/components/jetpack/licensing-activation';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
 import { addQueryArgs } from 'calypso/lib/url';
 import { isJetpackStandaloneProduct } from 'calypso/my-sites/plans/jetpack-plans/is-jetpack-standalone-product';
 import slugToSelectorProduct from 'calypso/my-sites/plans/jetpack-plans/slug-to-selector-product';
-import { SelectorProduct } from 'calypso/my-sites/plans/jetpack-plans/types';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
+import JetpackActivationInstructions from './jetpack-activation-instructions';
+import JetpackStandaloneActivationInstructions from './jetpack-standalone-activation-instructions';
 
 interface Props {
 	productSlug: string | 'no_product';
 	receiptId: number;
 }
-
-interface JetpackStandaloneActivationInstructionsProps {
-	product: SelectorProduct;
-}
-
-const JetpackPluginActivationInstructions: FC = () => {
-	const translate = useTranslate();
-
-	return (
-		<>
-			<p>{ translate( "If you don't have Jetpack installed, follow these instructions:" ) }</p>
-
-			<ul className="licensing-thank-you-manual-activation-instructions__list">
-				<li className="licensing-thank-you-manual-activation-instructions__list-item">
-					<span className="licensing-thank-you-manual-activation-instructions__step-number">1</span>
-					<span>
-						{ ' ' }
-						{ translate(
-							'Go to your WP Admin Dashboard and {{strong}}add a new plugin{{/strong}}.',
-							{
-								components: { strong: <strong /> },
-							}
-						) }
-					</span>
-				</li>
-
-				<li className="licensing-thank-you-manual-activation-instructions__list-item">
-					<span className="licensing-thank-you-manual-activation-instructions__step-number">2</span>
-					<span>
-						{ translate( 'Search for {{strong}}Jetpack{{/strong}}, install and activate.', {
-							components: { strong: <strong /> },
-						} ) }
-					</span>
-				</li>
-			</ul>
-
-			<p>
-				<ExternalLink
-					href="https://jetpack.com/support/install-jetpack-and-connect-your-new-plan/"
-					icon
-				>
-					{ translate( 'Learn more about how to install Jetpack' ) }
-				</ExternalLink>
-			</p>
-		</>
-	);
-};
-
-const JetpackStandaloneActivationInstructions: FC<
-	JetpackStandaloneActivationInstructionsProps
-> = ( { product } ) => {
-	const translate = useTranslate();
-	return (
-		<>
-			<ul className="licensing-thank-you-manual-activation-instructions__list">
-				<li className="licensing-thank-you-manual-activation-instructions__list-item">
-					<span className="licensing-thank-you-manual-activation-instructions__step-number">1</span>
-					<span>{ translate( 'Login to an existing Wordpress site as an administrator.' ) }</span>
-				</li>
-
-				<li className="licensing-thank-you-manual-activation-instructions__list-item">
-					<span className="licensing-thank-you-manual-activation-instructions__step-number">2</span>
-					<span>
-						{ translate(
-							'Go to {{strong}}Plugins > add New{{/strong}} in the admin menu on the left hand side.',
-							{
-								components: { strong: <strong /> },
-							}
-						) }
-					</span>
-				</li>
-
-				<li className="licensing-thank-you-manual-activation-instructions__list-item">
-					<span className="licensing-thank-you-manual-activation-instructions__step-number">3</span>
-					<span>
-						{ translate(
-							'Search for {{strong}}Jetpack %(pluginName)s{{/strong}}, install and activate.',
-							{
-								components: { strong: <strong /> },
-								args: { pluginName: product.shortName },
-							}
-						) }
-					</span>
-				</li>
-			</ul>
-
-			<p>
-				<ExternalLink
-					href="https://jetpack.com/support/install-jetpack-and-connect-your-new-plan/"
-					icon
-				>
-					{ translate( 'Learn more about how to install Jetpack %(pluginName)s', {
-						args: { pluginName: product.shortName },
-					} ) }
-				</ExternalLink>
-			</p>
-		</>
-	);
-};
 
 const LicensingActivationInstructions: FC< Props > = ( { productSlug, receiptId } ) => {
 	const translate = useTranslate();
@@ -172,7 +73,7 @@ const LicensingActivationInstructions: FC< Props > = ( { productSlug, receiptId 
 				{ jetpackStandaloneProduct ? (
 					<JetpackStandaloneActivationInstructions product={ jetpackStandaloneProduct } />
 				) : (
-					<JetpackPluginActivationInstructions />
+					<JetpackActivationInstructions />
 				) }
 
 				<Button
